moving stdout/err reading to Process
